Moved all setup logic into abstract parent class
